refactor(data/matrix/basic): work around leanprover/lean4#2042
This adjust definitions such that everything is well-behaved in the case that things are unfolded.
For each such definition, a lemma is added that replaces the equation lemma.

While this just adds `_apply` noise in Lean 3, it is necessary when porting to Lean 4 as there the equation lemma is not generated in the way that we want.