fix(topology/continuous_function/algebra): let `continuous_map.module'` use a `semiring` instead of ` ring` (#18868)
This fixes an issue where Lean couldn't recognize via TC inference that, for a topological semiring `R`, that `C(R, R)` is an `R`-module, despite the fact that it knows it is an `R`-algebra through `continuous_map.algebra`. All that is necessary is to weaken the type class assumptions.