refactor(category_theory/structured_arrow): make `structure_arrow` reducible
This doesn't look like it was intended as type synonym like `mul_opposite`; instead, it's just a shorthand like `module.dual`.

Unless these shorthands are reducible, they are not matched by Lean 4's discrimination trees in simp.
This is a problem because we mix an match the `structure_arrow` and `comma` API; which is itself an indication that the APIs are intended to be convenience shorthands for one another.